Move common test data to a `dummy` object
